"9am. Though it was 9:30pm, it was too early to go to bed at that time. Well, it is fine, but my rest was not great. Right now I feel groggy. Let me do my morning reading.
9:30am. Ok, one more chapter and then I will start. I will dedicate myself to dealing with packages. I am still thinking about how to do it. I thought of maybe making a supervisor that deals with multiple package servers, but nevermind that. I do not need concurrency for dealing with packages. I'll make that thing single threaded and have a single server manage all of them. That will make dealing with recursive checks easier.

9:35am. No, I like my concurrency. There is no need to lower it here.

9:50am. I am thinking about it. Let me turn off the router as I can't stop lurking Elaina threads.

9:55am. I have the whole day ahead of me. I have plenty of time to think before I start and that is what I am going to do.

10:05am. Ahhh, this desire.

A switch has been flipped in me. I keep thinking about the stock market and the knowledge that I finally have the ingredients I need to beat it are giving me focus.

Just what were those days of hardship for? In the end it all feels so simple.

There is no need for regrets. There is only desire in the end. It either drives you or it does not. Everything else is just foam on the waves.

10:15am. Right now, I feel like Livermore finally understanding what it means to sit tight. Even if I read his book, holding a position for over 3 months against market downturns would have been unthinkable to the past me.

I keep thinking about `TAN`. Sure I am bullish on it, but it is not like I expect it to go up 50% right away in the next 3 months. Even if it is going to double a few times, it is going to be a wild ride over the next 2 years. And I'd always need to be on guard in case the momentum is waning and this initial impulse turns out to be a dud. Given how stretched the market is, there will inevitably be shakeouts. Maybe `TAN` or `ARKK` will turn out surprisingly weak on market corrections and will force me take losses.

Either way, right now I want that hardship. Previously I only wanted to fit the market into a tight, understandable box. I wanted stocks to have the junk bond fund kind of persistent momentum. But now I want to stand against the waves, because I have the confidence I would be able to handle them.

10:25am. I want a fight.

Right now I can't trade, but what is a few years here and there to somebody taking the long term view?

10:30am. I am almost desire this as much as the Singularity right now. In the end, the markets won't matter, but this will of taking the long road will.

Really, the hardest thing in the world is what I've done so far - work every day for the past 5.5 years. And before that I spend so much time writing Simulacrum.

I see this as qualitatively different from my trading days when I only looked at charts and read the news. During the trading days, I fought to maintain a balance between hardness and laziness.

In contrast to this, holding a position and sitting on my ass would be nothing.

10:35am. It is not just stocks, holding a position that I will be able to make my agents and that I will be able to reach the Singularity is easy. The hard part is maintain that belief in the face of work.

Who wants to sit in this room constantly being faced with failure? Who wants to work for untold months on a language? Who wants to keep working on those agents that keep going nowhere?

Who wants to be looked down on for having ambition?

10:40am. But now most of my reasons for wanting something like money have died. The only thing I want it is because to some extent I need it, and because I want to prove that it is a triviality it really is. Peasants live their whole lives scrounging for such pitiful amounts. I want to go beyond that at least.

10:45am. It might be a triviality, but you need to look beyond the short term in order to be able to trully claim that.

And once you have that perspective, maybe having that perspective for its own sake is what really matters.

10:50am. That view really is power itself.

As I rant, I am thinking about the package server in the background.

I am pitting individual servers + a supervisor against having a single server that handles all of the files. The tug is going back and forth.

I think I know what scheme I want to use. Let me review the benefits and disadvantages of various approaches so I can explain my decision.

1) Individual servers + a supervisor.

This was the first thing I started thinking about. It is a straightforward extension of the current scheme. Right now I already have the individual server. So I thought of making a special supervisor that checks for circularity.

The benefits of this is that it maximizes concurrency in the system.

Of course having high concurrency means having low control. That is the disadvantage. I am afraid that having circularity checking outside the individual packages might make compilation more difficult - it is possible to have concurrency bugs when the user triggers the build, but the individual servers are still working on their own thing.

So I do not really like this.

2) Single server.

This would get rid of all the synchronization issues. This would even be easy to implement.

The issue is that every package would need to be loaded in sequential order. Right now my CPU is 4 cores so this is not a big deal, but I want to optimize the start up times. The place where doing things sequentially will hurt the most is right there.

So I have a proposal how to make this work.

I'll load the packages in a breadth first fashion.

Suppose I have packages
A - Q W E
B - R T Y
C - Z X V

If I were loading the packages sequentially, that would be in the depth first fashion. A, then is packages Q W E, then B, then R T Y, then C, then Z X and lastly V.

There is a lot of potential for concurrency here. A B C are independent of one another, and so are their submodules. It would be good not to waste this opportunity.

Rather than strictly aiming to maximize concurrency, what I can do is first deal with A B C. Then I can deal with Q W E R T Y Z X V.

I can just start A B C as separate jobs. It would be a list of jobs. Then I'd just use ... `Array.mapJob` in order to map this to a single result. Or use `Array.iterJobIgnore` and just send a message to the server.

This would be really good.

12:20pm. I am brainstorming like crazy here. I figured out how to do loading fully concurrently within a single server.

12:30pm. Let me stop here.

12:35pm. I need to go through this brainstorming session and after that I'll be done."