"9am. I am up. Let me chill a little. I really did go to bed at 6pm yesterday. I woke up during the night and my back was killing me due to being in bed for so long.
I did get some design insight though.

Previously, when a package has been modified, I thought I would just mutate its entry points.

I mean, the way I imagined the servers for Spiral would work is that they would have streams coming in and out, and in order to mutate those I would add control nodes which would add as relays.

Right now I am thinking that those things are trouble. It is really easy to screw this kind of thing up and introduce bugs. The reason why I decided to go against individual servers in favor of that big package machine is because it would be really hard to terminate a server if the user adds a package file.

Since I now understand the wonders of reverse graphs, I am going to ditch the control nodes entirely and instead rebuild everything downstream from the change.

9:10am. Also as I thought during the night, I realized that using a big state machine to manage typechecking for a package is really, really hard. It was one thing when it was just prototyping, but the full v0.2 has folders I need to deal with. Because of that not only do I have to go back and forth, I'd also need to manage scope as the machine goes upwards and outwards from a folder.

I get stress just thinking about it.

Instead of making that complicated machine, I am just going to compile the package to individual servers, but they are going to be owned by the package manager. Just screw anything else. It is much easier for me do a compilation pass than make a state machine for this kind of thing.

When the package changes I am just going terminate the unused servers and start afresh.

That is the way things should be. When a change is made, I should rebuild the downstream, not mutate it. Both the typechecking servers and package servers will have termination switches which will allow me to get rid of the old ones and implement the rebuilding scheme.

Yes, this is more work, since if a package is modified I will have to traverse its reverse graph and propagate the changes (or errors) throughout, but this is also much easier to reason about that single point mutations.

9:40am. Let me have breakfast here. This thing will be done when it wants to be.

Unlike yesterday, my mind does not feel split in a 100 different directions so I feel like I can take it easier today. I want to have some fun. After that I'll slowly start this thing. Forget finishing this by the end of the month. It will take a while.

The important thing is that I have the will to do it.

In that respect, yesterday's chaos mediation was a great success."