"9:50am. Yesterday I went to bed at 12pm.
https://community.acer.com/en/discussion/449171/aspire-e-15-wifi-dropping

This crappy issue on my bro's laptop is part of the reason.

Both my father and I haven't had any luck in dealing with this. When the guy brought me the laptop he said that the connection kept disconnecting, but I myself can't get it to turn on at all. I can't find anything on the internet to fix it so I am thinking of alternatives. As it turns out USB stick adapters exist and they go for 5-15$. Sure they would take an USB slot, but that is a reasonable tradeoff in this situation.

I want to avoid the situation where my brother is forced to do a refund, so I will make this suggestion.

10am. Hmmmm, my father still wants to fiddle with it. Well so be it. He can have it - there are some choices like trying to install Windows 7 in case the newer OS updates are doing something strange, but at that point I would be trying out random things in hopes of a miracle. I have better things to do with my time.

10:10am. Let me chill. Besides this laptop distraction I've been in a mood last night.

The fact that I can literally feel right now, in the programming part of my brain that I can beat the market is rousing my spirits. This isn't just idle boasting. Look at a junk bond fund chart, and you will see them having very persistent uptrends and downtrends on the daily scale.

The stock market is similar except on the monthly scale. When a particular group is in a strong uptrend, it very rarely has a significant negative month. It is all either up or flat. You make some allowances for when the general market is correcting. If it does make a negative month, then you can hypothesise that that the uptrend is broken. You get out and wait a month - if the market retraces the losses, you get back in.

Towards the end of my trading journey years ago, I actually hadn't seen this.

Like for example, I read Darvas and O'Neil's books - and they all use a price sensitive trend following system. Cutting your loss if a stock falls 7-10% is the notable feature of such systems. Together with what I put here from elsewhere, this make me obsessed about stocks making new highs. I spent years tracking that kind of system and getting nowhere with it.

10:25am. At first I was very inexperienced, but one thing throughout the six year journey that I could never figure out is why general trend following fails.

By that I mean - even though the stock market has a strong upward bias, I could never get around the fact that buying new high and selling lows is actually a poor system in the stock market. Especially if you are trading something like stock indices.

Even towards the end of the journey when the accumulated evidence was large that I really should be ignoring the daily price swings and thinking in terms of two week moves, I could not really accept that there is no good way of avoiding market corrections. I mean, there were people like Gary Smith who after 20 years of failure went on to very rarely having losing months so why couldn't I?

Before the 00s, it does seem that the markets had a strong short term persistent momentum. You can see it on some illiquid instrumets even today. The traders of those days knew to take advantage of that.

10:35am. But ultimately it does not matter. I can finally see that now. It does not matter that I cannot avoid corrections these days. It does not matter that I will have losing months or have to withstand long periods where the market goes nowhere.

New high and new lows do not matter, I can finally state that now. Daily non-persistence does not matter.

What really matters is the dance you do with the market. By making simple assumptions about price movement 3 months in advance it is possible to fade the daily noise.

10:45am. The rhythm you are trading in really does matter.

I mean, it would be beyond my imagination that once you buy you should hold for a month - if the stock goes down, shouldn't you cut it regardless of how long you held it? That is good trading isn't it according to all the books? And if it goes up you should get more bullish and buy it back.

It is all just noise. The infromation price gives you is also dependent on time. None of the books I've read talked about time.

It is really hard to see. I can only really see it because I took a step back and threw in the towel. When you are in the market - you can't be a budha. The urge is to improve and to fight. You never want to accept luck. You want to impose your own reality and stuggle.

10:50am. Well, let me put that rant aside. I really should be focusing on programming, but even before that let me chill. I want to do my morning reading and have breakfast. No point in starting anything now.

I've been thinking about concurrency in general. I have some thought about where I am going wrong and how to resolve the struggle, but I'll leave that rant for the follow up entries.

10:55am. I've been thinking about my thinking when it comes to this latest challenge, and definitely have some holes in my approach."