"9am. https://fundresearch.fidelity.com/fund-screener/
So it turns out Fidelity has an excellent screener for mutual funds, not just for its own family, but across the board. I've been trying to find this for days. I've been trying to talk myself into ditching individual stocks, but I was not sure that ETFs on their own would give me a wide enough breadth of options. With this, that issue has been settled.

Well, it is not true that ETFs are not broad enough in their offerings. That might have been true one or two decades ago, but today things are different. It is true that there are some offerings that the two do not intersect with, here I am just fulfilling a psychological need.

9:10am. Even if it is going to be a decent way into the future that it is going to be useful, I will commit myself to this new way of thinking today. If you don't want to spent time doing it, actually trading a basket is not easy for a lone trader - it is much easier to go through an intermediary.

Back when I was starting out, it was my dream that trading should not take more than 15-30m of my time a day. If I internalize this new way of thinking, that will be a success.

9:20am. Let me read Mahoako and Hero's Daughter, then I will get to work. I need to deal with the mess that are the validator and individual file servers. I might make the validator a regular function since it does not need to be concurrent.

9:25am. This recent rally, and the action of the past 6.5 years that I've been away from market really did fire me up. The fact that I'll be able to make peace with my position in the financial world is something that I did not expect. My effort was just way too much. Had I given up quicker, I'd have come to the answer sooner, but then again, if I were the type to give up, I'd never have attained the needed experience.

9:30am. The same goes for programming.

Let me do my morning reading so I can do my morning programming. I am wasting too much time.

10:15am. Let me start.

10:30am. I am thinking.

Basically everything is a huge mess. By everything, I really mean the way errors are handled.

I have the individual file server, and it is doing validation, but then I am also making and validating in the validator.

To make matters worse, the validator is supposed to be propagating errors, and yet it does not have the capability to see most of the errors because they are being handled in isolation in the file server. I am doing the same function in multiple places.

10:40am. Ah, yes, I nearly forgot. If the individual file server fails to load, it will raise a fatal error instead of just showing up as a regular one. That is nasty.

11:30am. I am thinking. I actually have a lot of inspiration right now. I have a great idea how to separate loading and validation. I figured out how to do concurrent depth first traversal.

12:30pm. Everything is just coming to me. Now only do I know how to do the parts that have been giving me trouble thus far, but I know how to do the entire compilation pipeline without any of the design complexities and disadvantages of my current approach.

12:50pm. For suppressing parser inputs to the typechecker, I do not have an elegant solution, but I can think of couple of rough ones. Well, this is just a small fragment of the overall thing, so I do not need to worry about this.

12:55pm. Yeah, I can do this. It all fits now. I really was short of inspiration, but with the ideas I have in mind now, I do not have to make a hugely complicated chain of state machines. For build itself, I can do a concurrent DFS much like for loading files and validating them.

Let me stop here. I am overdue for breakfast."