"3:50pm. Yeah, I don't think I have anything else I want to do before starting work on the next step.
Let me step away from the screen for half an hour so I can think about it. It is time to move to the next stage. I've dreamed about this for a long time and it will finally be time to do it.

While I am doing this, I will also take the change to go back and make the nominals and prototypes have a package id as a part of their identity. I said that a flat number would not do the trick. It needs to be a `package_id, tag` pair

I'll leave that for last though.

3:55pm. Since this will be so complex, I am going to stick to my principles strongly and try to not get tangled up in the details. Instead I will loosen up and just focus on what a single function should be doing and not make too many assumptions about anything else.

Ultimately, there are very real limits to how much I can understand about the code I am writing myself. I need to be careful not to cross them, and modularizing and abstracting according to those limits is what is good programming.

```fs
let change (schemas : Dictionary<_,_>) links loads errors is_open dir text =
    match schemas.TryGetValue(dir) with
    | true, Ok _ when is_open -> Job.unit()
    | _ ->
        schemas.Remove(dir) |> ignore
```

Don't need this last line as `validate` will do that anyway. Let me get rid of it.

```fs
let load (m : LoadedSchemas) text project_dir =
    let m = Map.remove project_dir m
```

Also let me remove the project dir in here.

This is the way to do it.

4:05pm. Let me just run it. I doubt anything will break, but I should not just assume that the changes will stick.

4:10pm. Everything works as expected. I just moved things around a little, but now things are more integruous. I do not have functions whose purpose should not be to do so removing stuff from the dictionaries and maps.

4:10pm. Let me finally step away from the screen for a bit."