"9am. Let me chill a bit and then I will start.
9:15am. Let me start. Focus me. I've started a second playthrough of BG, but I really should just move to BG2 or something else. Still, the game is just a magical as I remember it. One of the best RPGs of all time, easily. When BG2 came out, I was surprised that it was actually even better.

Now that I am older, I see that the conversations all follow a certain structure where you have the polite and rude choices and the later will just piss of people and get you worse rewards.

So even if you are trying to play Chaotic Evil, all the Good choices always give you better rewards. If you are on the side of Good, the rewards are supposed to just manifest if you affect the right attitude. The designers of the game really do not understand Evil.

9:20am. Ultimately, their understanding of it is limited from the framework of Good. If I ever make my own game, it will definitely have an outside perspective.

Suppose BG was made from an Evil perspective - you'd be able to do game breaking stuff like rob shopkeepers of all their loot, and murder and robbery in general would be the superior option. You are already a murderhobo, but the best stuff is always concentrated in the civilized lands.

One thing the Reverend Insanity got right is that since cultivation in that story depends on resources, that is a great motivator for Fang Yuan to take the demonic path.

Once you level up, what are the people you've harmed going to do? They won't be able to touch you. That is the Evil perspective.

9:25am. In the real world, I do not think either Evil or Good are the dominant frameworks - rather as we approach the Singularity, there is only your own effort.

As for myself, the people around me cannot help me here, but they sure could delay me if I gave them the opportunity. So that is what the peasants amount to in the end, the era of mass produced weaponry has given them a swelled head. There is no point in harming them, but there is no point in being with them. This neutrality is loneliness and aggressive avoidance. It is being an expert at running away.

I do not owe them anything. My true friends are away from me. Shared goals, shared resources, shared power - that is the essence of friendship, what do you call those people who are friends on paper, but those criteria do not apply to? Acquaintances.

My own's effort is the real truth in life, and to unlock it I needed the vision of the suicide-loop self improvement. That vision is the truth of Singularity. All I need to do is make steady effort to get closer to it. Maybe I won't get to it, maybe I will, either way I will have done the right thing in life.

9:35am. The world I want is the one I will have to create.

So let me do my part for today.

Let me make that file server. Yesterday I did not want to start at 5pm, but now I have the whole day ahead of me. I need to get back to par.

```fs
type FileReq =
    | FOpen of {|uri : string; spiText : string|}
    | FChanged of {|uri : string; spiEdit : SpiEdit|}
    | FTokenRange of {|uri : string; range : VSCRange|} * VSCTokenArray IVar
    | FHoverAt of {|uri : string; pos : VSCPos|} * string option IVar
```

Let me start with this. Now I have a clear target.

10:10am. I am getting lost in thought.

```fs
let file_server fatal_errors tokenizer_errors parser_errors type_errors req req_aux =
    let dict = Dictionary()
    let fresh (uri : string) = {|
        tokenizer=tokenizer
        parser=parser (System.IO.Path.GetExtension(uri) = ".spi")
        typechecker=typechecker 0 0 Infer.top_env_default
        |}
    let get (uri : string) = Utils.memoize dict fresh uri
    let req = Stream.values req
    let req_aux = Stream.values req_aux

    let rec read =
        req ^=> function
            | FOpen x ->
                match dict.TryGetValue x.uri with
                | true, x -> ()
                | _ -> fresh.tokenizer.Run(DocumentAll(x.spiText))
    ()
```

Yeah, this is not going like this. But I have a fairly simple vision of how to maximally simplify the server.

Let me do that file stream after all. Once I do that, that will open the door for extensions to it.

11:45am. I am still thinking about it.

12:40pm. i am still thinking about it. I am in the flow state; I am having a lot of inspiration.

I think I understand what the problem was with the old servers. Or rather, I've realized that they have a problem and my new idea is superior.

```fs
type TypecheckerRes = (Bundle * Infer.InferResult * Infer.TopEnv) PersistentVector * bool
let typechecker package_id module_id (req : ParserRes Stream) : TypecheckerRes Stream =
    let req = Stream.values req
    let res = Src.create()
    let r = Src.tap res
    let rec waiting a = req ^=> fun b ->
        let rec loop s i b =
            match PersistentVector.tryNth i a, b with
            | Some(bundle,_,_ as r), bundle' :: b' when bundle = bundle' -> loop (PersistentVector.conj r s) (i+1) b'
            | _ -> s, b
        loop PersistentVector.empty 0 b.bundles |> processing
    and processing = function
        | a, [] -> Alt.prepare (Src.value res (a,true) >>- fun () -> waiting a)
        | a, b :: b' -> waiting a <|> Alt.prepare (Src.value res (a, false) >>- fun () ->
            let env =
                match PersistentVector.tryLast a with
                | Some(_,_,env) -> env
                | None -> Infer.top_env_default
            let r = Infer.infer package_id module_id env (bundle_top b)
            let a' = PersistentVector.conj (b,r,Infer.union r.top_env_additions env) a
            processing (a', b')
            )
    Hopac.server (waiting PersistentVector.empty)
    r
```

This is the old thing. The state transitions are very explicit here, but if I had just starting the typechecker as a separate process + a cancellation token, I could have made all of this a straightforward loop.

1:05pm. Damn it, I know how to make everything much better, but there is a small sync issue when it comes to error reporting that I can't work out. This is another one of those annoying 99% of the time it will work fine, but the other 1% it will be wrong, but won't affect the compilation anyway, yet I still have to work it out. It is that kind of annoying bug. The problem with cancellation tokens is that in the lag between the old process ending and the new process starting, the errors get out of line.

Even putting an if statement directly on `Src.value errors` would not deal with it 100%.

2:05pm. Done with breakfast. Let me go do the chores. I think I know how to deal with the issue that has been confusing me. I need to change my mindset. Since I obviously need to wait for the job to finish, instead of starting it and forgetting about it, in addition to the cancellation token, I'll keep around a promise of that job. Then I'll just wait on that promise when it is time to go to the next loop iteration in the server.

I think this situation is finally a good use for `Promise.start`. I finally found a reason to use it. I am really in awe at the foresight of the library's author. He really did a good job. Hopac is FParsec tier in its thoroughness.

An S ranked library all the way."