"5:25pm. I am back.
```fs
let supervisor tokenizer_errors parser_errors type_errors package_errors fatal_errors req =
    let loop = req >>= function
        | ProjectFileOpen x -> failwith "TODO"
        | ProjectFileChange x -> failwith "TODO"
        | ProjectFileDelete x -> failwith "TODO"
        | ProjectFileLinks(x,res) -> failwith "TODO"
        | ProjectCodeActions(x,res) -> failwith "TODO"
        | ProjectCodeActionExecute x -> failwith "TODO"
        | FileOpen x -> failwith "TODO"
        | FileChanged x -> failwith "TODO"
        | FileTokenRange(x, res) -> failwith "TODO"
        | HoverAt(x,res) -> failwith "TODO"
    Job.foreverServer loop
```

My focus when I am away from the screen has not really been good. When I do the chores, my mind starts to wander towards other things.

It is not like when I am taking a nap specifically to refine my ideas.

I am weak. In order to really focus I need to sit down and grind away at it.

5:30pm. Let me review my old code for a while. I need to run my desire over and over in order to build it up. I need to manifest the urge.

5:50pm. I've been thinking of various things.

Yeah, it is telling that of all the things I could think about, the way the `file_server` was done is what stuck in my mind. I can't use cancellation tokens. That general technique is too complicated.

In a similar vein...

```fs
    let rec loop old_results =
        {new TypecheckerStream with
            member _.Run(res) =
                let r =
                    top_env >>=* fun top_env ->
                    res >>= fun res ->
                    // Doing the memoization like this has the disadvantage of always focing the evaluation of the previous
                    // streams' first item, but on the plus side will reuse old state.
                    old_results >>- fun old_results ->
                    run (cons_fulfilled old_results) top_env 0 res.bundles
                let a = Stream.mapFun (fun (_,x,_) -> x) r
                a, loop r
            }
```

I've been thinking about the way I've done this part. I dislike this.

I want it the memoization to be done here similarly to how it is done in the parser. This way is no good as it does excess computation because it always forces the first item and will create garbage in the scenarios I am envisioning.

The parser's way is the best tradeoff. Even if it will potentially throw aways useful old state, it won't create garbage.

I can in fact redesign this to use the same pattern. Today I'll just add a TODO, but tomorrow I'll move `old_results >>- fun old_results ->` up and see where it leads me from there.

In the past few months I've done a lot of thinking about concurrency. And one thing that is clear to me now that I am bathed in the stresses of real world needs is that an amateur thinks of increasing concurrency in order to get the benefits of it.

Now that I am closer to an expert, I see that to get the benefits I should be instead increasing laziness. Instead of launching more processes, make things immutable and do less work.

And things being pure and lazy will make it quite natural to uncover good concurrency opportunities. Like with packages.

Apart from missing some pieces, the way it is done now is very good. I have multiple promises and in order to unlock the benefit of concurrency there, all I have to do is grab them in parallel.

Right now the package streams return: `Map<string, InferResult Stream> * PackageDiffStream`.

Both the core and the diff stream return that map of `InferResult Stream`s.

I am thinking of adding another layer of laziness to that. Later on in addition to this map of infer results, I'll need to ...

Actually, I am pointing at the wrong thing here.

6:05pm. I am confused now.

```fs
type PackageCoreStateItem = {
    links : Map<string,{|name : string|}>
    rev_links : string Set
    top_env_in : TopEnv Promise
    top_env_out : TopEnv Promise
    stream : MultiFileStream
    }
```

I meant to point at the `top_env_in`. The way I am imagining it, I need to diff the hierarchy and only then can I pass the top env to the next package.

But instead of doing everything in `add_package` eagerly I'd like to expose even more laziness just in case the user working on some package that links to many other packages. I do not want every single key press to immediatelly affect all the packages linking to it.

```fs
packages : Map<string,PackageCoreStateItem>
```

I am not sure whether I should aim to lazify this further at this particular spot, but I'll definitely consider replacing `PackageCoreStateItem` with `PackageCoreStateItem Promise` here. I'll leave it as a future idea.

6:10pm. It is actually surprising that I do not want anything to do with cancellation tokens. I spent a lot of time thinking of how to deal with launching and cancelling concurrent processes, but now that time has passed and I am looking at it again, I feel completely cold on it.

6:15pm. Now that the push has come to shove, I am only really composing promises and streams. I do not even want to think about how to make use of that technique. I can't see any place where I need it either.

So I guess it will be fine to leave it aside.

...

The strangest thing right now is just how well things are clicking for me.

I've been frazzled the whole day, and you'd think that nothing would come to me givem my emotional state, but I can't see any difficulties at all from my present vantage point.

6:30pm. I am thinking.

Right now, if the user loads a package file, the server will in fact try to recursively load the packages it links to. I need to go beyond that. I need to have the loading the package file also load the modules.

It does not go one way only.

Most of the time, a module file is the one that will get opened first, so what the program will have to do is look for the project file, load it and then load the rest of the modules.

So modules must be capable of loading packages and packages must be capable of loading modules.

So far I've only been relying on VS Code to ferry the file data around, but I'll have to redesign the server to be more proactive.

I am thinking how exactly do I want to do this loading.

I do not want the schema validator to say no problem only to get an error later when I try to load the file from disk. So that functionality is going to have to be close to there.

Schema validation is almost an ancient piece by now, but I am going to have to reimagine its purpose and extend its functionality.

So far for simplicity I've been taking for granted that the files are loaded in memory, but now I am going to have to decide a few things.

Let me make a little TODO list.

1) Fix the way packages are merged in `add_package`.
2) Lazify the `typechecker` more like the `parser`.
3) Extend the package loader so it loads modules as well.

Lunch time.

6:55pm. I am back, but obviously I am not going to be doing any more programming at this time. It is time for some fun.

It is strange, my BG obsession just does not seem to be abbating. Well, I'll scratch that itch at some point.

I want to see what exactly is the ideal game."