"9:35am. I remember as a kid once I finished a game and it started getting boring, I'd try out various cheats. It never made it fun again.
With BG is similar, except I am so obsessed with it now that this is what is making it fun.

I really have no idea what to think about Spell Revisions.

9:55am. I am explode if I keep thinking about it for any longer. But with the latest changes I do not think solo play is viable - at least it would not be if not for that basilisk zone that gives you >25k free exp.

10am. You know, maybe the real question I should be asking myself is - in settings with magic, what is the point in even having fighters and other non-magic using classes?

When I get down to it, BG is not actually that good at magic - the best game that I've had the pleasure of experiencing with a mage is Dungeon Crawl Stone Soup.

In BG, you can walk until you see the enemy, cast fireball, and by the time you finish casting the melee guys have already caught up to you. In DCSS you can cast it a few times and kill him before he walks up to you and then scram before his friends get there. In BG to get the same effect, you have to use invis and hurl fireballs from out of line of sight.

I want to dig deeper into it, but is there a point in playing when all I am doing is AI abuse?

Right now I seem to be stuck in a loop of replaying the early parts of BG without being able to find a release.

10:05am. Abuse is right, but better AI is also right.

So the ideal game is the one that continually gets better along with you.

I really like Spell Revision the idea of buffing summons, but I do not know how to feel about it nerfing sleep so heavily and making Blindness a level 8 spell with a much shorter duration.

10:15am. Let me chill for a while. I do not feel like starting right away.

My sleep last night was not that good, I want to purge my BG thoughts out of my system. I've more or less tried everything but playing a thief by now.

10:20am. The way to purge these thoughts is to decide what my hopes are. When I started playing BG, I really wanted to master mages that I did not as a kid. But now...I'll put it on a backburner until I decide. Just this kind of thinking is a source of fun for me now.

11:15am. It is good that I did not start as the breakfast was early.

Right now my obsession is at its peak, so let me go outside for a bit and do the chores for a while. The weather cleared up finally.

11:20am. All this tripping down the memory lane has reminded me of a certain idea I had as a kid. It might be worth cultivating at this point. I'll think about it.

1:35pm. Done with chores. Less than a week of winter preparation left.

Let me cool down for a while.

2pm. Every day is a first step.

Let me make one. Today I will focus solely on module loading.

It is time to elaborate it.

2:45pm. I have some ideas and new principles in my mind, but before I talk about it, let me step away from the screen for half an hour. I want to sleep on this for a while and want to do it away from the screen.

After I do that, I'll write down my ideas and start work on them.

I'll start with the typechecker lazification and then redo the entire schema loading and validation process. I am going to change how it works. I'll get rid of that concurrent loading function as it makes everything non-composable.

Let me commit here."